refactor: move base64 shared kernel to 'shared' top-level package
