+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/configclient/domain/SpringCloudConfigModuleFactory.java
@@ -11,13 +11,13 @@ import static tech.jhipster.lite.module.domain.JHipsterModule.propertyValue;
 import static tech.jhipster.lite.module.domain.JHipsterModule.springProfile;
 import static tech.jhipster.lite.module.domain.JHipsterModule.toSrcMainDocker;
 
-import tech.jhipster.lite.generator.base64.domain.Base64Utils;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.docker.DockerImages;
 import tech.jhipster.lite.module.domain.file.JHipsterSource;
 import tech.jhipster.lite.module.domain.javaproperties.JHipsterModuleSpringProperties.JHipsterModuleSpringPropertiesBuilder;
 import tech.jhipster.lite.module.domain.javaproperties.PropertyValue;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
+import tech.jhipster.lite.shared.base64.domain.Base64Utils;
 import tech.jhipster.lite.shared.error.domain.Assert;
 
 public class SpringCloudConfigModuleFactory {
