+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/eureka/domain/EurekaModuleFactory.java
@@ -10,12 +10,12 @@ import static tech.jhipster.lite.module.domain.JHipsterModule.propertyValue;
 import static tech.jhipster.lite.module.domain.JHipsterModule.to;
 import static tech.jhipster.lite.module.domain.JHipsterModule.versionSlug;
 
-import tech.jhipster.lite.generator.base64.domain.Base64Utils;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.docker.DockerImages;
 import tech.jhipster.lite.module.domain.file.JHipsterSource;
 import tech.jhipster.lite.module.domain.javaproperties.PropertyValue;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
+import tech.jhipster.lite.shared.base64.domain.Base64Utils;
 import tech.jhipster.lite.shared.error.domain.Assert;
 
 public class EurekaModuleFactory {
